refactor: 121-move isFetchError function into utils/type-guards
