+++ src/utils/type-guards/is-fetch-error.ts
@@ -1,4 +1,4 @@
-import type { FetchError } from '../types';
+import type { FetchError } from '../../app/types';
 import type { ProductProjectionPagedQueryResponse } from '@commercetools/platform-sdk';
 
 export function isFetchError(data: ProductProjectionPagedQueryResponse | FetchError): data is FetchError {
